ux: ミーティング詳細のボタンラベルと削除ボタン配置を改善する (issue #111)
- 「ミーティングを開始」→「記録を開始」に変更し、記録モード開始の動作を明確化
- 削除ボタンを独立した赤いボタンから「その他（...）」ドロップダウン内に移動し誤操作を防止
- MeetingHeaderActions コンポーネントを新規作成（DropdownMenu + AlertDialog）
- MeetingDeleteDialog を MeetingHeaderActions に統合して削除
